feat: refactors function to utils
