+++ swimlane/core/adapters/record.py
@@ -5,7 +5,7 @@ from swimlane.core.cache import check_cache
 from swimlane.core.resolver import AppResolver
 from swimlane.core.resources.record import Record, record_factory
 from swimlane.core.resources.report import Report
-from swimlane.utils import random_string, one_of_keyword_only
+from swimlane.utils import random_string, one_of_keyword_only, validate_type
 from swimlane.utils.version import requires_swimlane_version
 
 
@@ -417,17 +417,3 @@ def validate_filters_or_records(filters_or_records):
             raise ValueError("Expected filter tuple or Record, received {0}".format(item))
 
     return _type
-
-def validate_type(field, value):
-    """Type validation for filters and fields from bulk_modify, bulk_delete and filter"""
-    accepted_values_to_check = (int, str, float, list, bool, tuple)
-    should_check_value_type = not value == None and type(value) in accepted_values_to_check
-    if should_check_value_type:
-        value_list = value if isinstance(value, list) else [value]
-        for v in value_list:
-            wrong_type = not any(
-                    [isinstance(v, field_type) for field_type in field.supported_types]
-            ) if len(field.supported_types) > 0 else False
-
-            if wrong_type:
-                raise ValueError("Value must be one of {}".format(", ".join([str(f) for f in field.supported_types])))
